backend/pyproject.toml
# backend/pyproject.toml
[tool.poetry]
name = "secure-cms"
version = "1.0.0"
description = "A secure contact management system"
authors = ["Secure CMS Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.109.2"
starlette = "0.36.3"
uvicorn = {extras = ["standard"], version = "0.27.1"}
gunicorn = "22.0.0"
sqlalchemy = "1.4.42"
alembic = "1.13.1"
psycopg2-binary = "2.9.9"
asyncpg = "0.29.0"
databases = {extras = ["postgresql"], version = "0.8.0"}
PyJWT = "2.8.0"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
python-multipart = "0.0.7"
pydantic = {extras = ["email"], version = "2.6.1"}
pydantic-settings = "2.1.0"
python-dotenv = "1.0.1"
pyotp = "2.9.0"
argon2-cffi = "23.1.0"
cryptography = "43.0.3"
httpx = "0.26.0"
requests = "2.32.3"
urllib3 = "2.2.3"
certifi = "2024.8.30"
aiohttp = "3.9.4"
jinja2 = "3.1.3"
aiofiles = "23.2.1"
python-magic = "0.4.27"
email-validator = "2.1.0.post1"
orjson = "3.9.15"
ujson = "5.9.0"
redis = "5.0.1"
celery = "5.3.6"
flower = "2.0.1"
prometheus-client = "0.19.0"
sentry-sdk = "1.39.1"
structlog = "24.1.0"
python-json-logger = "2.0.7"
boto3 = "1.34.34"
pillow = "10.2.0"
packaging = ">=22.0"
bcrypt = "4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "8.0.0"
pytest-asyncio = "0.23.5"
pytest-cov = "4.1.0"
pytest-xdist = "3.5.0"
coverage = "7.4.1"
black = "24.1.1"
flake8 = "7.0.0"
pylint = "3.0.3"
mypy = "1.8.0"
bandit = "1.7.7"
pip-audit = "2.7.2"
ipython = "8.21.0"
ipdb = "0.13.13"
watchfiles = "0.21.0"
isort = "5.13.2"
pytest-env = "1.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
